Add FormatMessage helper method
This primarily moves code from Win32Exception to be publicly accessible. It also adds tests.
