+++ src/Kernel32.Tests.Shared/Kernel32.Tests.Shared.projitems
@@ -11,6 +11,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Kernel32.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Win32ExceptionTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)KernelExtensionsTests.cs" />
   </ItemGroup>
 </Project>
